+++ scripts/common/__workdir-exec.sh
@@ -789,14 +789,13 @@ workdir_exec() {
     fi
   fi
 
-  # For now we support delete of localnet only.
-  # Need to be more careful for testnet/devnet to preserve key.
   if [ "$CMD_DELETE_REQ" = true ]; then
     stop_all_services
 
     if ! $is_local; then
+      # Do more precise deletion for testnet/devnet/mainnet to preserve key.
       (if cd "$SUI_REPO_DIR" >&/dev/null; then cargo clean; fi)
-      rm -rf "$WORKDIRS/$WORKDIR/logs/sui.log/*"
+      rm -rf "$WORKDIRS/$WORKDIR/logs/sui.log"
       rm -rf "$WORKDIRS/$WORKDIR/config/sui-process.log"
       rm -rf "$WORKDIRS/$WORKDIR/config/sui-faucet-process.log"
       if [ -d "$WORKDIRS/$WORKDIR/sui-repo-default" ]; then