+++ typescript/vscode-extension/CHANGELOG.md
@@ -1,12 +1,12 @@
 # Change Log
 
-All notable changes to the "suibase" extension will be documented in this file.
+When upgrading the extension, do also "~/suibase/update" to get the latest Suibase CLI.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## 0.1.6
+  - Make installation/upgrade more user friendly, better error messages.
 
 ## 0.1.5
   - Various fix related to installation.
-  - Do "~/suibase/update" to get the latest Suibase CLI.
 
 ## 0.1.4
   - Warn user if some Sui prerequisites are not installed (git and rustc)